Use Integer with Partitions

+++ symmetric_group/character.h
@@ -80,17 +80,16 @@ struct SymmetricGroupCharacter {
     return *this;
   }
 
-  friend double InnerProduct(SymmetricGroupCharacter const &lhs,
-                             SymmetricGroupCharacter const &rhs) {
+  friend Integer InnerProduct(SymmetricGroupCharacter const &lhs,
+                              SymmetricGroupCharacter const &rhs) {
     if (lhs.values_.empty()) { return 0; }
-    int64_t result = 0;
+    Integer result = 0;
     for (auto const &[partition, coefficient] : lhs.values_) {
       auto iter = rhs.values_.find(partition);
       if (iter == rhs.values_.end()) { continue; }
       result += (iter->second * coefficient * CycleTypeCount(partition));
     }
-    return static_cast<double>(result) /
-           Factorial(lhs.values_.begin()->first.whole());
+    return result / Factorial(lhs.values_.begin()->first.whole());
   }
 
   friend std::ostream &operator<<(std::ostream &os,
